+++ src/test/java/be/ugent/idlab/knows/mappingweaver/spec/rml_kgc/RMLCoreTest.java
@@ -1,13 +1,17 @@
 package be.ugent.idlab.knows.mappingweaver.spec.rml_kgc;
 
-import be.ugent.idlab.knows.mappingweaver.cores.TestCore;
+import java.util.stream.Stream;
+
 import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.parallel.Execution;
+import org.junit.jupiter.api.parallel.ExecutionMode;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import java.util.stream.Stream;
+import be.ugent.idlab.knows.mappingweaver.cores.TestCore;
 
+@Execution(ExecutionMode.CONCURRENT)
 public class RMLCoreTest extends TestCore {
     private static Stream<Arguments> unfixable() {
         return Stream.of(